Add HTTP tests
Nice to be able to test end-to-end that the backend does what's expected.

Tests work by getting a TCP port from the OS. This allows us to run
multiple tests in parallel.

We add tests for starting the server and trying to add a booking when
not logged in.

The database isn't isolated or reset. Support is needed before adding
tests that mutate the database or depend on its state.

There is no bypassing of the log-in system in place. This means that we
can't test anything that depends on being logged in.

There is no time-faking support. This limits our ability of testing
things that would take time to complete. The tests would take the amount
of wall-clock time required for the operation. We don't want slow tests
so time-faking would be needed before such tests are added.