Only allow observation for booked user
User needs to have booked the specific telescope to be allowed to
observe. We can also do some cleanup with duplicate checks for active
bookings (that didn't take telescope into account).

Maybe it would be nicer if you would be redirected (say if you have a
window open when the booking runs out). This just fixes the immediate
bug where the wrong user would have access to the observe UI of a
telescope.

This closes #198.