Make telescope resource locking internal
Make the update function an implementation detail of the telescope and
move all locking into the telescope implementation modules. This will
facilitate removing the TelescopeHandle and simplifying the program
structure.

This also means that it is possible to mutate a telescope through a
non-mut reference or binding. This is fine since there isn't really
value semantics per telescope anyways (there is really only one of each,
cloning doesn't create a new physical telescope!).